+++ src/sass/_overview.scss
@@ -1,3 +1,5 @@
+@import "./helpers";
+
 .overview-content {
   $golden-ratio: 1.618;
 
@@ -12,10 +14,22 @@
 
   .overview-card {
     .p-card__title {
-      i {
-        height: 1.5rem;
-        margin-bottom: calc(-1 * $spv--small / 2);
-        width: 1.5rem;
+      align-items: center;
+      display: flex;
+      justify-content: space-between;
+      max-width: unset;
+      width: 100%;
+
+      .overview-card-title {
+        i {
+          height: 1.5rem;
+          margin-bottom: calc(-1 * $spv--small / 2);
+          width: 1.5rem;
+        }
+      }
+
+      .explanation-tooltip-wrapper {
+        margin-top: calc(-0.45rem - 24px);
       }
     }
 
@@ -95,6 +109,37 @@
 
       .plus-n-more {
         margin-left: $sph--small;
+      }
+    }
+
+    &.warnings {
+      .warnings-table-overview-scroll {
+        max-height: 17.5rem;
+        overflow-y: auto;
+
+        .warnings-table--overview {
+          thead th {
+            background-color: var(--vf-color-background-default);
+            position: sticky;
+            top: 0;
+            z-index: 1;
+          }
+
+          thead .last_message {
+            display: table-cell;
+            overflow: visible;
+            white-space: nowrap;
+          }
+
+          tbody .last_message {
+            @include u-line-clamp-truncate(2);
+
+            /* Explicitly cap the height so overflow lines can't bleed through */
+            max-height: calc(
+              1.5rem * 2 + 0.5rem
+            ); // line-height * lines + padding
+          }
+        }
       }
     }
   }
